+++ L04-qa/src/main/java/ru/otus/testing/example1/services/WeddingService.java
@@ -1,0 +1,7 @@
+package ru.otus.testing.example1.services;
+
+import ru.otus.testing.example1.model.Family;
+
+public interface WeddingService {
+    Family doWedding();
+}
