+++ L04-qa/src/main/java/ru/otus/testing/example2/services/ClosedConsoleIOService.java
