+++ L04-qa/src/main/java/ru/otus/testing/exercise/MessageBuilder.java
