+++ L04-qa/src/main/java/ru/otus/testing/exercise/TemplateNotFoundException.java
