+++ L04-qa/src/test/java/ru/otus/testing/example2/services/ClosedIOServiceTest.java
