--- L05-qa/src/main/java/ru/otus/testing/example1/services/HumanGenerator.java
@@ -1,7 +1,0 @@
-package ru.otus.testing.example1.services;
-
-import ru.otus.testing.example1.model.Human;
-
-public interface HumanGenerator {
-    Human generate(char sex);
-}
